Map Dataproc JSON path semantics by Spark shim
